requirements.txt
discord.py==0.16.0
aiohttp==1.0.5
wikipedia==1.4.0
urbandictionary==1.1
PyNaCl==1.1.2